requirements.txt
aws-cdk-lib==2.146.0
constructs>=10,<11
pytest==8.3.2
pytest-cov==5.0.0
boto3==1.34.158
